requirements.txt
pandas==1.5.3
pytest==7.2.2
PyYAML==6.0
